+++ libft/ft_memcmp.c
@@ -3,16 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memcmp.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fran <fran@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: frangome <frangome@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/21 11:32:52 by fran              #+#    #+#             */
-/*   Updated: 2023/04/22 17:53:22 by fran             ###   ########.fr       */
+/*   Updated: 2023/04/26 18:06:44 by frangome         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_memcmp(const void *s1, const void *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned int	counter;
+	size_t	counter;
 
 	counter = 0;
 	while (counter < n)
