+++ libft/ft_memcpy.c
@@ -3,16 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memcpy.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fran <fran@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: frangome <frangome@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/20 17:59:19 by frangome          #+#    #+#             */
-/*   Updated: 2023/04/22 17:59:09 by fran             ###   ########.fr       */
+/*   Updated: 2023/04/26 18:07:06 by frangome         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dest, const void *src, unsigned int n)
+#include "libft.h"
+
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	unsigned int	c;
+	size_t	c;
 
 	c = 0;
 	if (!dest && !src)
