+++ libft/ft_strlcpy.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcpy.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fran <fran@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: frangome <frangome@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/18 17:50:52 by frangome          #+#    #+#             */
-/*   Updated: 2023/04/25 21:16:20 by fran             ###   ########.fr       */
+/*   Updated: 2023/04/26 18:10:40 by frangome         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlcpy(char *dest, char *src, unsigned int size)
+size_t	ft_strlcpy(char *dest, char *src, size_t size)
 {
-	unsigned int	count;
-	unsigned long	i;
+	size_t	count;
+	size_t	i;
 
 	i = 0;
 	while (src[i] != 0)
