+++ libft/libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fran <fran@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: frangome <frangome@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/17 15:52:24 by frangome          #+#    #+#             */
-/*   Updated: 2023/04/25 21:17:55 by fran             ###   ########.fr       */
+/*   Updated: 2023/04/26 18:13:13 by frangome         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,24 +27,24 @@ int		ft_isalpha(int c);
 int		ft_isascii(int c);
 int		ft_isdigit(int digit);
 int		ft_isprint(int c);
-void	*ft_memset(void *str, int c, unsigned int n);
+void	*ft_memset(void *str, int c, size_t n);
 char	*ft_strchr(const char *str, int c);
-int		ft_strlen(const char *str);
-int		ft_strncmp(const char *s1, const char *s2, unsigned int n);
+size_t	ft_strlen(const char *str);
+int		ft_strncmp(const char *s1, const char *s2, size_t n);
 char	*ft_strrchr(const char *str, int c);
 int		ft_tolower(int upper);
 int		ft_toupper(int lower);
-void	ft_bzero(void *s, unsigned int n);
-void	*ft_memchr(const void *s, int c, unsigned int n);
-size_t	ft_strlcat(char *dest, const char *src, unsigned int size);
-size_t	ft_strlcpy(char *dest, char *src, unsigned int size);
-char	*ft_strnstr(const char *big, const char *lit, unsigned int l);
+void	ft_bzero(void *s, size_t n);
+void	*ft_memchr(const void *s, int c, size_t n);
+size_t	ft_strlcat(char *dest, const char *src, size_t size);
+size_t	ft_strlcpy(char *dest, char *src, size_t size);
+char	*ft_strnstr(const char *big, const char *lit, size_t l);
 int		ft_atoi(const char *str);
-void	*ft_calloc(unsigned int nmemb, unsigned int size);
+void	*ft_calloc(unsigned int nmemb, size_t size);
 char	*ft_strdup(const char *s);
-void	*ft_memcpy(void *dest, const void *src, unsigned int n);
-int		ft_memcmp(const void *s1, const void *s2, unsigned int n);
-void	*ft_memmove(void *dest, void *src, unsigned int n);
+void	*ft_memcpy(void *dest, const void *src, size_t n);
+int		ft_memcmp(const void *s1, const void *s2, size_t n);
+void	*ft_memmove(void *dest, void *src, size_t n);
 char	*ft_substr(char const *s, unsigned int start, size_t len);
 char	*ft_strjoin(char const *s1, char const *s2);
 char	*ft_strtrim(char const *s1, char const *set);
